fgsm-service/requirements.txt
torch
torchvision
fastapi
uvicorn
pillow
requests
